basic-critc+agent/pyproject.toml
[project]
name = "summarizer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.16.0",
]
